*** CLOUDS don't work right now! 	You cannot currently have a cloud as a child of a preset. 	So the child needs to be instantiated in the parent's instantiation, 	not in the preset.
This currently wrecks the river/beach placers.

===

RichSoil changes (Laterite in jungle, peat in swamps) also got added here
(opps -- this is supposed to be the water placer branch).

1. Swamps in vanilla are now more varied in height changes. There is significant
   "above sea" / dry zones, *AND* also significant "below sea" / wet zones.

The current behavior is not idea. Will check with Noogenesis; I suspect that
the better behavior is to only target "shovel" blocks below sea level, and to
increase the frequency in the process. I'm not sure if the "above sea" zones should
be valid targets for bog iron / peat-deposits. My understanding of real life
is that above ground wetlands have too much drainage for this type of ore deposit.

Issue: Deposits were too low to be found in swamps.
Main swamp/bog deposit moved from 55 +/- 5 (so it can center down at Y=51, and go
   down to Y=48, and then tilt down another 2-3) to Y=58.5 +/- 4 "normal" - a goal
   of Y=56 to 60 for the center, which should show in underwater swamps, especially
   with tilt.
In addition to dirt and stone, clay and gravel were added to the targets. (NB: *Sand*?)
   Again, replacement made 30% ore, 40% clay, 30% unchanged.

BUG!
Laterite deposits (jungles/mod biomes) DO show up deep below the surface layers.
   So that does need to be made deep -- the height change needs to be moved to the peat
   distribution, not the parent Rich Soil distribution.


Cloud Height for Harder-PresetLaterite needs to be made to scale to biome once that is fixed.
Heights for Harder-PresetPeat needs to scale to sea level.